extract pods table
