+++ cyclops-ui/src/components/k8s-resources/Deployment.tsx
@@ -1,21 +1,8 @@
 import React, { useEffect, useState } from "react";
-import {
-  Button,
-  Col,
-  Divider,
-  Row,
-  Table,
-  Alert,
-  Tag,
-  Tabs,
-  Modal,
-  TabsProps,
-} from "antd";
+import { Col, Divider, Row, Alert } from "antd";
 import axios from "axios";
-import { formatPodAge } from "../../utils/pods";
-// import { DownloadOutlined } from "@ant-design/icons";
-import ReactAce from "react-ace";
 import { mapResponseError } from "../../utils/api/errors";
+import PodTable from "./common/PodTable";
 
 interface Props {
   name: string;
@@ -26,14 +13,6 @@ const Deployment = ({ name, namespace }: Props) => {
   const [deployment, setDeployment] = useState({
     status: "",
     pods: [],
-  });
-  const [logs, setLogs] = useState("");
-  const [logsModal, setLogsModal] = useState({
-    on: false,
-    namespace: "",
-    pod: "",
-    containers: [],
-    initContainers: [],
   });
   const [error, setError] = useState({
     message: "",
@@ -67,120 +46,6 @@ const Deployment = ({ name, namespace }: Props) => {
     };
   }, []);
 
-  const handleCancelLogs = () => {
-    setLogsModal({
-      on: false,
-      namespace: "",
-      pod: "",
-      containers: [],
-      initContainers: [],
-    });
-    setLogs("");
-  };
-
-  const downloadLogs = (container: string) => {
-    return function () {
-      window.location.href =
-        "/api/resources/pods/" +
-        logsModal.namespace +
-        "/" +
-        logsModal.pod +
-        "/" +
-        container +
-        "/logs/download";
-    };
-  };
-
-  const getTabItems = () => {
-    let items: TabsProps["items"] = [];
-
-    let container: any;
-
-    if (logsModal.containers !== null) {
-      for (container of logsModal.containers) {
-        items.push({
-          key: container.name,
-          label: container.name,
-          children: (
-            <Col>
-              <Button
-                type="primary"
-                // icon={<DownloadOutlined />}
-                onClick={downloadLogs(container.name)}
-              >
-                Download
-              </Button>
-              <Divider style={{ marginTop: "16px", marginBottom: "16px" }} />
-              <ReactAce
-                style={{ width: "100%" }}
-                mode={"sass"}
-                value={logs}
-                readOnly={true}
-              />
-            </Col>
-          ),
-        });
-      }
-    }
-
-    if (logsModal.initContainers !== null) {
-      for (container of logsModal.initContainers) {
-        items.push({
-          key: container.name,
-          label: "(init container) " + container.name,
-          children: (
-            <Col>
-              <Button
-                type="primary"
-                // icon={<DownloadOutlined />}
-                onClick={downloadLogs(container.name)}
-              >
-                Download
-              </Button>
-              <Divider style={{ marginTop: "16px", marginBottom: "16px" }} />
-              <ReactAce
-                style={{ width: "100%" }}
-                mode={"sass"}
-                value={logs}
-                readOnly={true}
-              />
-            </Col>
-          ),
-        });
-      }
-    }
-
-    return items;
-  };
-
-  const onLogsTabsChange = (container: string) => {
-    axios
-      .get(
-        "/api/resources/pods/" +
-          logsModal.namespace +
-          "/" +
-          logsModal.pod +
-          "/" +
-          container +
-          "/logs",
-      )
-      .then((res) => {
-        if (res.data) {
-          let log = "";
-          res.data.forEach((s: string) => {
-            log += s;
-            log += "\n";
-          });
-          setLogs(log);
-        } else {
-          setLogs("No logs available");
-        }
-      })
-      .catch((error) => {
-        setError(mapResponseError(error));
-      });
-  };
-
   return (
     <div>
       {error.message.length !== 0 && (
@@ -207,107 +72,9 @@ const Deployment = ({ name, namespace }: Props) => {
           Replicas: {deployment.pods.length}
         </Divider>
         <Col span={24} style={{ overflowX: "auto" }}>
-          <Table dataSource={deployment.pods}>
-            <Table.Column
-              title="Name"
-              dataIndex="name"
-              filterSearch={true}
-              key="name"
-            />
-            <Table.Column title="Node" dataIndex="node" />
-            <Table.Column title="Phase" dataIndex="podPhase" />
-            <Table.Column
-              title="Started"
-              dataIndex="started"
-              render={(value) => <span>{formatPodAge(value)}</span>}
-            />
-            <Table.Column
-              title="Images"
-              dataIndex="containers"
-              key="containers"
-              width="15%"
-              render={(containers, record: any) => (
-                <>
-                  {containers.map((container: any) => {
-                    let color = container.status.running ? "green" : "red";
-
-                    if (record.podPhase === "Pending") {
-                      color = "yellow";
-                    }
-
-                    return (
-                      <Tag
-                        color={color}
-                        key={container.image}
-                        style={{ fontSize: "100%" }}
-                      >
-                        {container.image}
-                      </Tag>
-                    );
-                  })}
-                </>
-              )}
-            />
-            <Table.Column
-              title="Logs"
-              width="15%"
-              render={(pod) => (
-                <>
-                  <Button
-                    onClick={function () {
-                      axios
-                        .get(
-                          "/api/resources/pods/" +
-                            namespace +
-                            "/" +
-                            pod.name +
-                            "/" +
-                            pod.containers[0].name +
-                            "/logs",
-                        )
-                        .then((res) => {
-                          if (res.data) {
-                            let log = "";
-                            res.data.forEach((s: string) => {
-                              log += s;
-                              log += "\n";
-                            });
-                            setLogs(log);
-                          } else {
-                            setLogs("No logs available");
-                          }
-                        })
-                        .catch((error) => {
-                          setError(mapResponseError(error));
-                        });
-                      setLogsModal({
-                        on: true,
-                        namespace: namespace,
-                        pod: pod.name,
-                        containers: pod.containers,
-                        initContainers: pod.initContainers,
-                      });
-                    }}
-                    block
-                  >
-                    View Logs
-                  </Button>
-                </>
-              )}
-            />
-          </Table>
+          <PodTable namespace={namespace} pods={deployment.pods} />
         </Col>
       </Row>
-      <Modal
-        title="Logs"
-        open={logsModal.on}
-        onOk={handleCancelLogs}
-        onCancel={handleCancelLogs}
-        cancelButtonProps={{ style: { display: "none" } }}
-        width={"60%"}
-      >
-        <Tabs items={getTabItems()} onChange={onLogsTabsChange} />
-      </Modal>
     </div>
   );
 };
